move Package.cs to Core/ folder
