+++ test/PackageTest.cs
@@ -1,4 +1,5 @@
-﻿using Netly.Core;
+﻿using Netly.Abstract;
+using Netly.Core;
 
 namespace test;
